[Fix] #245 - API & Service 변경
