+++ iOS-NOTTODO/iOS-NOTTODO/Network/Service/Recommend/RecommendService.swift
@@ -9,41 +9,75 @@ import Foundation
 
 import Moya
 
-enum RecommendService {
-    case recommend
-    case action(id: Int)
-    case situdation
+typealias RecommendData = GeneralArrayResponse<RecommendResponseDTO>
+typealias ActionData = GeneralResponse<RecommendActionResponseDTO>
+typealias SituationData = GeneralArrayResponse<RecommendSituationResponseDTO>
+
+protocol RecommendServiceType {
+    func getRecommend(completion: @escaping (RecommendData?) -> Void)
+    func getRecommendAction(index: Int, completion: @escaping (ActionData?) -> Void)
+    func getRecommendSituation(completion: @escaping (SituationData?) -> Void)
 }
 
-extension RecommendService: BaseService {
-    var domain: BaseDomain {
-        return .recommend
-    }
+final class RecommendService: RecommendServiceType {
     
-    var urlPath: String {
-        switch self {
-        case .recommend:
-            return URLConstant.recommend
-        case .action(id: let id):
-            return URLConstant.recommendAction + "/\(id)" + "/action"
-        case .situdation:
-            return URLConstant.recommendSituation
+    static let shared: RecommendService = RecommendService()
+    
+    private let provider = MoyaProvider<RecommendAPI>(session: Session(interceptor: AuthInterceptor.shared), plugins: [MoyaLoggingPlugin()])
+    
+    private init() {}
+        
+    // MARK: - GET
+    
+    func getRecommend(completion: @escaping (RecommendData?) -> Void) {
+        provider.request(.recommend) { result in
+            switch result {
+            case .success(let response):
+                do {
+                    let response = try response.map(RecommendData?.self)
+                    completion(response)
+                } catch let err {
+                    print(err.localizedDescription, 500)
+                }
+            case .failure(let err):
+                print(err.localizedDescription)
+                completion(nil)
+            }
         }
     }
     
-    var headerType: HeaderType {
-        return .jsonWithToken
+    func getRecommendAction(index: Int, completion: @escaping (GeneralResponse<RecommendActionResponseDTO>?) -> Void) {
+        provider.request(.action(id: index)) { result in
+            switch result {
+            case .success(let response):
+                do {
+                    let reponse = try response.map(ActionData?.self)
+                    completion(reponse)
+                } catch let err {
+                    print(err.localizedDescription, 500)
+                }
+            case .failure(let err):
+                print(err.localizedDescription)
+                completion(nil)
+            }
+        }
     }
     
-    var method: Moya.Method {
-        return .get
-    }
-    
-    var validationType: ValidationType {
-        return .successCodes
-    }
-    
-    var task: Moya.Task {
-        .requestPlain
+    func getRecommendSituation(completion: @escaping (SituationData?) -> Void) {
+        provider.request(.situdation) { result in
+            switch result {
+            case .success(let response):
+                do {
+                    let response = try
+                    response.map(SituationData?.self)
+                    completion(response)
+                } catch let err {
+                    print(err.localizedDescription, 500)
+                }
+            case .failure(let err):
+                print(err.localizedDescription)
+                completion(nil)
+            }
+        }
     }
 }
