+++ iOS-NOTTODO/iOS-NOTTODO/Presentation/AddMission/Cells/NottodoCollectionViewCell.swift
@@ -178,7 +178,7 @@ extension NottodoCollectionViewCell {
     }
     
     private func requestRecentMissionAPI() {
-        MissionAPI.shared.getRecentMission { [weak self] response in
+        MissionService.shared.getRecentMission { [weak self] response in
             guard self != nil else { return }
             guard let response = response else { return }
             guard let data = response.data else { return }
