+++ iOS-NOTTODO/iOS-NOTTODO/Presentation/AddMission/Cells/SituationCollectionViewCell.swift
@@ -175,7 +175,7 @@ private extension SituationCollectionViewCell {
     }
     
     func requestRecommendSituationAPI() {
-        RecommendAPI.shared.getRecommendSituation { [weak self] response in
+        RecommendService.shared.getRecommendSituation { [weak self] response in
             guard self != nil else { return }
             guard let response = response else { return }
             guard let data = response.data else { return }
