+++ iOS-NOTTODO/iOS-NOTTODO/Presentation/AddMission/ViewControllers/AddMissionViewController.swift
@@ -280,7 +280,7 @@ extension AddMissionViewController {
     private func requestPostAddMission(title: String, situation: String,
                                        actions: [String]?, goal: String?, dates: [String]?) {
         let request = AddMissionRequest(title: title, situation: situation, actions: actions, goal: goal, dates: dates ?? [""])
-        MissionAPI.shared.postAddMission(request: request) { response in
+        MissionService.shared.postAddMission(request: request) { response in
             guard let response = response else { return }
             switch response.status {
             case 200..<300:
@@ -304,7 +304,7 @@ extension AddMissionViewController {
     
     private func requestPutUpdateMission(id: Int, title: String, situation: String, actions: [String]?, goal: String?) {
         let request = UpdateMissionRequest(id: id, title: title, situation: situation, actions: actions, goal: goal)
-        MissionAPI.shared.putUpdateMission(request: request) { response in
+        MissionService.shared.putUpdateMission(request: request) { response in
             guard let response = response else { return }
             print(response.status)
             switch response.status {
@@ -327,7 +327,7 @@ extension AddMissionViewController {
     }
     
     private func requestGetMissionDates(id: Int) {
-        MissionAPI.shared.particularMissionDates(id: id) { [weak self] response in
+        MissionService.shared.particularMissionDates(id: id) { [weak self] response in
             guard let data = response.data else { return }
             for item in data {
                 self?.dateList.append(item)
@@ -337,7 +337,7 @@ extension AddMissionViewController {
     }
     
     private func requestDailyMissionAPI(id: Int) {
-        MissionAPI.shared.getDetailMission(id: id) { [weak self] response in
+        MissionService.shared.getDetailMission(id: id) { [weak self] response in
             guard let self = self, let response = response else { return }
             
             if let data = response.data {
