+++ iOS-NOTTODO/iOS-NOTTODO/Presentation/Home/ViewControllers/MissionDetailViewController.swift
@@ -227,7 +227,7 @@ extension MissionDetailViewController {
 extension MissionDetailViewController {
     
     func requestDailyMissionAPI(id: Int) {
-        MissionAPI.shared.getDetailMission(id: id) { [weak self] response in
+        MissionService.shared.getDetailMission(id: id) { [weak self] response in
             guard let self else { return }
             guard let response = response else { return }
             guard let data = response.data else { return }
@@ -243,7 +243,7 @@ extension MissionDetailViewController {
     }
     
     private func requestDeleteMission(id: Int) {
-        MissionAPI.shared.deleteMission(id: id) { [weak self] _ in
+        MissionService.shared.deleteMission(id: id) { [weak self] _ in
             guard let self else { return }
             guard let data = self.detailModel else { return }
             if data.id == id {
